Remove all self-referencing to module and direct to Prelude
New game plan: going to push all the references outward to Prelude
so these only act as a dictionary of bindings for rahse. Then all the
modules remain totally distinct from each other, and can be subbed in
and out at will, adding or removing bindings so long as the call graph
within the module is complete.