Remove all capitalized names
Classes, newtypes, GADTs and datatype decls are all removed, with the
necessary `import` statement to import them instead from Prelude.
This is done messily with several regex hacks, but it gets most of the
way there. We also create files of the instance function definitions so
these can be imported qualified in `Union`